+++ .gitignore
@@ -29,14 +29,14 @@ build_jaxlib/dist
 
 # Examples build and tmp files
 examples/build/
-examples/opt_serving/dataset/*.so
-examples/opt_serving/dataset/*.c
-examples/opt_serving/dataset/*.cpp
-examples/opt_serving/weblogs
-examples/opt_serving/benchmark/tmp*
-examples/opt_serving/tmp*
+examples/llm_serving/dataset/*.so
+examples/llm_serving/dataset/*.c
+examples/llm_serving/dataset/*.cpp
+examples/llm_serving/weblogs
+examples/llm_serving/benchmark/tmp*
+examples/llm_serving/tmp*
 examples/opt_finetune/output/
-examples/language_model/norwegian-gpt2/
+examples/gpt2/norwegian-gpt2/
 alpa_debug_info
 
 # Analysis temp files
