+++ examples/llm_serving/model/test_cache.py
@@ -4,7 +4,7 @@ import jax.numpy as jnp
 import numpy as np
 
 from alpa.testing import assert_allclose
-from opt_serving.model.opt_model import (get_opt_config, init_model_aval,
+from llm_serving.model.opt_model import (get_opt_config, init_model_aval,
                                          inference_step_no_cache,
                                          init_cache_np,
                                          build_position_ids,
