+++ examples/llm_serving/scripts/step_2_consolidate_992_shards_to_singleton.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from tqdm import tqdm
 
 import torch
-from opt_serving.scripts.utils import load_and_pop_last_optimizer_state
+from llm_serving.scripts.utils import load_and_pop_last_optimizer_state
 
 logger = logging.getLogger(__name__)
