+++ examples/llm_serving/scripts/step_3_convert_to_numpy_weights.py
@@ -5,7 +5,7 @@ import argparse
 import os
 
 import numpy as np
-from opt_serving.scripts.utils import torch_load_cpu
+from llm_serving.scripts.utils import torch_load_cpu
 
 
 def save_numpy(weight_dict, to_folder):
