+++ examples/llm_serving/service/img.png
