+++ examples/llm_serving/service/index.html
@@ -365,7 +365,7 @@
 
       <div class="d-grid gap-4 d-sm-flex justify-content-sm-center">
         <a href="#generation" class="btn btn-primary px-4 btn-lg">Try Live Generation</a>
-        <a href="https://alpa-projects.github.io/tutorials/opt_serving.html" class="btn btn-outline-primary px-4 btn-lg" target="_blank">Host Your Own Service</a>
+        <a href="https://alpa-projects.github.io/tutorials/llm_serving.html" class="btn btn-outline-primary px-4 btn-lg" target="_blank">Host Your Own Service</a>
       </div>
   </div>
 
@@ -571,8 +571,8 @@
                 <a href="https://huggingface.co/docs/transformers/v4.20.1/en/main_classes/text_generation#transformers.generation_utils.GenerationMixin.generate" target="_blank">a diverse set of generation techniques and arguments</a>.
             </p>
               <p>We are developing a RESTFUL API to expose the full set of arguments. Stay tuned.
-              Meanwhile, if you want to try out different generation techniques and hyperparameters now, you can <a href="https://alpa-projects.github.io/tutorials/opt_serving.html" target="_blank">set up your own OPT-175B service using Alpa</a>
-                  and start from <a href="https://github.com/alpa-projects/alpa/blob/main/examples/opt_serving/benchmark/benchmark_text_gen.py#L183" target="_blank">here</a>.</p>
+              Meanwhile, if you want to try out different generation techniques and hyperparameters now, you can <a href="https://alpa-projects.github.io/tutorials/llm_serving.html" target="_blank">set up your own OPT-175B service using Alpa</a>
+                  and start from <a href="https://github.com/alpa-projects/alpa/blob/main/examples/llm_serving/benchmark/benchmark_text_gen.py#L183" target="_blank">here</a>.</p>
           </div>
         </div>
       </div>
