+++ examples/setup.py
@@ -1,5 +1,5 @@
 import sys
 from setuptools import find_packages, setup
 
-setup(name="opt_serving",
+setup(name="llm_serving",
       packages=find_packages())
